pyproject.toml
[tool.poetry]
name = "meeting-scheduler"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
pygments = "^2.18.0"
psycopg = "^3.2.1"
httpie = "^3.2.3"
bcrypt = "^4.2.0"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
django-cors-headers = "^4.4.0"
python-dateutil = "^2.9.0.post0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
